commit adds functionality to Sponsor for change password and change email. Updated the add sponsor functionality

+++ src/main/java/com/cpsc4910/cpscbackend/Sponsor/SponsorRespository.java
@@ -2,5 +2,5 @@ package com.cpsc4910.cpscbackend.Sponsor;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface SponsorRespository extends CrudRepository<Sponsor, Integer> {
+public interface SponsorRespository extends CrudRepository<Sponsor, Long> {
 }
